requirements.txt
torch>=2.0
torchvision>=0.15
timm>=0.6.12
numpy
pandas
scikit-learn
tqdm
Pillow
matplotlib
